README.md
# Basera
Basera is an online marketplace for arranging lodging, primarily homestays, or tourism experiences across its users build on ANDROID STUDIO.


Idea behind the application:-
The idea of the application is to make an accommodation-sharing Android Application that
will allow people to Host their free-living space (Rooms or house) for others to stay.
Letting complete strangers into your home might not sound like a good idea to
everyone, neither is letting yourself into the home of a complete stranger, but with such
kind of app that give ratings to users and keeps track of their previous sharing or staying
experiences, it will not stays the same. Already Airbnb is a giant in the market trusted by
millions of users.

There are few questions that will come to the mind of everyone after knowing the basic
idea. How is it going to work, actually? Why people will use it?

Coming to the first question, the working of the application is very basis. The app will
allow people to rent out their properties or spare rooms to guests, from a shared room
to an entire house anything can be rented and made available to guests.
The second question, well Hosts can cash in on spare space, means that if someone have
some free space available, they will definitely want to earn some money from it, who
wouldn’t. Secondly, some hosts would just enjoy meeting new people and showing off
their home 😉.
Once you have booked, you and the host will arrange where and when to meet. The app
is going to provide contact information and other necessary instruction for the meeting
to both the guests and the hosts.